Added tests to AirService
