+++ src/Services/Air/AirService.js
@@ -26,7 +26,7 @@ module.exports = function (settings) {
       `${templatesDir}/AIR_PRICE_REQ.xml`,
       'air:airPriceRsp',
       AirValidator.FARE_RULES_TRIPS_TRAVELER_REFS,
-      null,
+      AirParser.AIR_ERRORS,
       AirParser.AIR_PRICE_REQUEST,
       debug
     ),
@@ -46,8 +46,9 @@ module.exports = function (settings) {
       `${templatesDir}/AIR_PRICE_REQ_MANUAL.xml`,
       'air:airPriceRsp',
       AirValidator.AIR_PRICE_MANUAL,
-      null,
-      AirParser.AIR_PRICE_REQUEST
+      AirParser.AIR_ERRORS,
+      AirParser.AIR_PRICE_REQUEST,
+      debug
     ),
     createReservation: uApiRequest(
       config.AirService.url,
@@ -85,8 +86,9 @@ module.exports = function (settings) {
       `${templatesDir}/AIR_PRICING_FARE_RULES.xml`,
       'air:airPriceRsp',
       AirValidator.FARE_RULES_BOOKED,
-      null,
-      AirParser.AIR_PRICE_FARE_RULES
+      AirParser.AIR_ERRORS,
+      AirParser.AIR_PRICE_FARE_RULES,
+      debug
     ),
     fareRulesTripsTravellerRefs: uApiRequest(
       config.AirService.url,
@@ -94,8 +96,9 @@ module.exports = function (settings) {
       `${templatesDir}/AIR_PRICING_FARE_RULES.xml`,
       'air:airPriceRsp',
       AirValidator.FARE_RULES_TRIPS_TRAVELER_REFS,
-      null,
-      AirParser.AIR_PRICE_FARE_RULES
+      AirParser.AIR_ERRORS,
+      AirParser.AIR_PRICE_FARE_RULES,
+      debug
     ),
     fareRulesUnbooked: uApiRequest(
       config.AirService.url,
@@ -103,8 +106,9 @@ module.exports = function (settings) {
       `${templatesDir}/AIR_PRICING_FARE_RULES.xml`,
       'air:airPriceRsp',
       AirValidator.FARE_RULES_BOOKED,
-      null,
-      AirParser.AIR_PRICE_FARE_RULES
+      AirParser.AIR_ERRORS,
+      AirParser.AIR_PRICE_FARE_RULES,
+      debug
     ),
     fareRulesUnbooked_uAPI: uApiRequest(
       config.AirService.url,
@@ -112,8 +116,9 @@ module.exports = function (settings) {
       `${templatesDir}/AIR_FARE_RULES_REQUEST.xml`,
       'air:AirFareRulesRsp',
       AirValidator.FARE_RULES_UAPI,
-      null,
-      AirParser.FARE_RULES_RESPONSE
+      AirParser.AIR_ERRORS,
+      AirParser.FARE_RULES_RESPONSE,
+      debug
     ),
     gdsQueue: uApiRequest(
       config.GdsQueueService.url,
